requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
hades-hash==0.1.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
